Create modules folder and inside this folder create function.js file for the functionality
