+++ src/modules/function.js
@@ -1,0 +1,36 @@
+let tasks = JSON.parse(localStorage.getItem('tasks')) || [];
+
+function deleteList() {
+  const button = document.querySelectorAll('.newButton');
+  const element = button[0].parentNode;
+  const liItem = element.parentNode;
+  liItem.removeChild(element);
+  const listId = element.getAttribute('id');
+  const index = parseInt(listId.substring(1), 10);
+  tasks = tasks.filter((task) => task.index !== index);
+  for (let i = 0; i < tasks.length; i += 1) {
+    tasks[i].index = i + 1;
+  }
+  localStorage.setItem('tasks', JSON.stringify(tasks));
+  window.location.reload();
+}
+
+export default function addTrash(event) {
+  const eventCatcher = event.target.closest('button');
+  eventCatcher.style.display = 'none';
+  const liItem = event.target.closest('li');
+  const nodeList = liItem.querySelectorAll('p');
+  const parentLi = nodeList[0].parentNode;
+  const parentId = parentLi.getAttribute('id');
+  const parentElement = document.getElementById(parentId);
+  const newBtn = document.createElement('button');
+  newBtn.className = 'newButton';
+  newBtn.innerHTML = '<i class="fa fa-trash-alt"></i>';
+  parentElement.append(newBtn);
+  liItem.style.backgroundColor = '#ffe6ee';
+  const textElem = liItem.querySelector('p');
+  textElem.setAttribute('contenteditable', true);
+  textElem.focus();
+  const trash = document.getElementsByClassName('newButton');
+  trash[0].addEventListener('click', deleteList);
+}
